o2-sim: More robust shutdown
This improves the shutdown sequence of o2-sim and tries to
avoid blocking situations in the termination phase
when waiting for all child processes to shut down.

This happened for instance during some igprof profiling
(for an unknown reason ... but most likely due to some signals that
were not delivered to all children).

The solution here is to install a final async timer thread which
will force-shutdown everything after some grace period.

+++ run/O2SimDeviceRunner.cxx
@@ -43,8 +43,9 @@ void sigaction_handler(int signal, siginfo_t* signal_info, void*)
   auto pid = getpid();
   LOG(info) << pid << " caught signal " << signal << " from source " << signal_info->si_pid;
   auto groupid = getpgrp();
-  if (pid == gMasterProcess) {
+  if (pid == gMasterProcess && signal_info->si_pid != gDriverProcess) {
     // master worker forwards signal to whole worker process group
+    // (do this only if not coming from gDriverProcess since this uses killpg and already affected all children)
     killpg(pid, signal);
   } else {
     if (signal_info->si_pid != gDriverProcess) {
